# [UWP] Use platform character conversion functions instead of codecvt