--- a/source/uwp/Renderer/lib/DateTimeParser.h
+++ b/source/uwp/Renderer/lib/DateTimeParser.h
@@ -3,7 +3,6 @@
 #pragma once
 #include "DateTimePreparser.h"
 
-#include <codecvt>
 #include <string>
 
 namespace AdaptiveNamespace