--- a/source/uwp/Renderer/lib/Util.cpp
+++ b/source/uwp/Renderer/lib/Util.cpp
@@ -1,8 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 #include "pch.h"
-#include <locale>
-#include <codecvt>
 #include <string>
 #include <regex>
 
@@ -63,25 +61,58 @@ HRESULT WStringToHString(const std::wstring& in, _Outptr_ HSTRING* out)
     return WindowsCreateString(in.c_str(), static_cast<UINT32>(in.length()), out);
 }
 
+std::string WstringToString(const std::wstring& in)
+{
+    if (!in.empty())
+    {
+        const size_t requiredSize =
+            WideCharToMultiByte(CP_UTF8, 0 /*dwFlags*/, in.c_str(), (int)in.length(), nullptr, 0, nullptr, nullptr);
+        std::string converted(requiredSize, 0);
+
+        if (WideCharToMultiByte(CP_UTF8, 0 /*dwFlags*/, in.c_str(), (int)in.length(), &converted[0], (int)requiredSize, nullptr, nullptr) == 0)
+        {
+            throw bad_string_conversion();
+        }
+        return std::move(converted);
+    }
+    return "";
+}
+
+std::wstring StringToWstring(const std::string& in)
+{
+    if (!in.empty())
+    {
+        // TODO: safer casts
+        const size_t requiredSize =
+            MultiByteToWideChar(CP_UTF8, 0 /*dwFlags*/, in.c_str(), (int)in.length(), (LPWSTR) nullptr, 0);
+        std::wstring wide(requiredSize, 0);
+
+        if (MultiByteToWideChar(CP_UTF8, 0 /*dwFlags*/, in.c_str(), (int)in.length(), &wide[0], (int)requiredSize) == 0)
+        {
+            throw bad_string_conversion();
+        }
+
+        return std::move(wide);
+    }
+    return L"";
+}
+
 HRESULT UTF8ToHString(const std::string& in, _Outptr_ HSTRING* out)
 {
     if (out == nullptr)
     {
         return E_INVALIDARG;
     }
-    std::wstring_convert<std::codecvt_utf8_utf16<wchar_t>> converter;
-    std::wstring wide = converter.from_bytes(in);
-    return WindowsCreateString(wide.c_str(), static_cast<UINT32>(wide.length()), out);
+    else
+    {
+        std::wstring wide = StringToWstring(in);
+        return WindowsCreateString(wide.c_str(), static_cast<UINT32>(wide.length()), out);
+    }
 }
 
 HRESULT HStringToUTF8(const HSTRING& in, std::string& out)
 {
-    if (in == nullptr)
-    {
-        return E_INVALIDARG;
-    }
-    std::wstring_convert<std::codecvt_utf8_utf16<wchar_t>> converter;
-    out = converter.to_bytes(WindowsGetStringRawBuffer(in, nullptr));
+    out = WstringToString(WindowsGetStringRawBuffer(in, nullptr));
     return S_OK;
 }
 
@@ -1207,8 +1238,9 @@ CATCH_RETURN;
 
 HRESULT StringToJsonObject(const std::string& inputString, _COM_Outptr_ IJsonObject** result)
 {
-    std::wstring asWstring = StringToWstring(inputString);
-    return HStringToJsonObject(HStringReference(asWstring.c_str()).Get(), result);
+    HString asHstring;
+    RETURN_IF_FAILED(UTF8ToHString(inputString, asHstring.GetAddressOf()));
+    return HStringToJsonObject(asHstring.Get(), result);
 }
 
 HRESULT HStringToJsonObject(const HSTRING& inputHString, _COM_Outptr_ IJsonObject** result)
@@ -1246,8 +1278,9 @@ HRESULT JsonObjectToHString(_In_ IJsonObject* inputJson, _Outptr_ HSTRING* resul
 
 HRESULT StringToJsonValue(const std::string inputString, _COM_Outptr_ IJsonValue** result)
 {
-    std::wstring asWstring = StringToWstring(inputString);
-    return HStringToJsonValue(HStringReference(asWstring.c_str()).Get(), result);
+    HString asHstring;
+    RETURN_IF_FAILED(UTF8ToHString(inputString, asHstring.GetAddressOf()));
+    return HStringToJsonValue(asHstring.Get(), result);
 }
 
 HRESULT HStringToJsonValue(const HSTRING& inputHString, _COM_Outptr_ IJsonValue** result)
@@ -1325,18 +1358,6 @@ HRESULT IsBackgroundImageValid(_In_ ABI::AdaptiveNamespace::IAdaptiveBackgroundI
     }
     *isValid = FALSE;
     return S_OK;
-}
-
-std::wstring StringToWstring(const std::string& in)
-{
-    std::wstring_convert<std::codecvt_utf8_utf16<wchar_t>, wchar_t> utfConverter;
-    return utfConverter.from_bytes(in);
-}
-
-std::string WstringToString(const std::wstring& input)
-{
-    std::wstring_convert<std::codecvt_utf8_utf16<wchar_t>, wchar_t> utfConverter;
-    return utfConverter.to_bytes(input);
 }
 
 void RemoteResourceElementToRemoteResourceInformationVector(_In_ ABI::AdaptiveNamespace::IAdaptiveElementWithRemoteResources* remoteResourceElement,