## Related Issue
Fixes #3398

## Description
The UWP renderer relies on the classes from stl's `codecvt` which have been deprecated. On top of that, `codecvt` is quite slow compared to platform APIs. With this change, we switch over to `MultiByteToWideChar()` and `WideCharToMultiByte()` for our charset conversion.

## How Verified
PerfApp and the test app. Between this change and the change that went in with #3379, we're seeing a pretty serious perf improvement:

### Before
![image](https://user-images.githubusercontent.com/16614499/63804403-f7092a00-c8cb-11e9-8f60-a82eab9716a8.png)

### After
![image](https://user-images.githubusercontent.com/16614499/63804411-fc667480-c8cb-11e9-9de0-de4aa983d901.png)

### Methodology

Measurements taken on my dev machine running an x64 release build of the PerfApp. The app was run without a debugger attached.

If you've never used the UWP PerfApp before, the above screenshots might be a little confusing. I chose 50 parse iterations and 1 render iteration per card. As implemented, this means that each card will be parsed and then rendered. This happens 50 times -- in other words, the number of parses and renders is equal (see [here](https://github.com/microsoft/AdaptiveCards/blob/ef004dbe74c58750b6d8da5577961510d2d828bd/source/uwp/PerfApp/MainPage.xaml.cpp#L95-L117)). I tested against the `source/samples/v1.0` folder.

"Ticks" according to PerfApp are the result of calling `GetTickCount64()` ([docs](https://docs.microsoft.com/en-us/windows/win32/api/sysinfoapi/nf-sysinfoapi-gettickcount64)). That is, each tick is a millisecond. The percentages listed after the average ticks is the percentage of the overall load time is taken by each process. Before our changes, cards took an average of 3ms to parse and they took 6ms to render, so 1/3 of the time was parse, and 2/3 spent on render. After our changes, parsing dropped (on average) to sub-millisecond times and rendering dropped to 2ms on average. You can see this improvement in the total render ticks as well, with the before figure being `27447` (27 seconds) and the after figure at `9208` (9 seconds).

###### Microsoft Reviewers: [Open in CodeFlow](http://wpcp.azurewebsites.net/CodeFlowProtocolProxy2.php?pullrequest=https://github.com/microsoft/AdaptiveCards/pull/3400)